Fix bug with absolute paths to Windows logs
This fixes the issue where Carbon will crash if using an absolute path
for `--log_path` on Windows.
